time_control: Limit time per move to a fraction of the remaining time.
Don't let the engine take too much time if there are lots
of moves remaining.

STC no regression
13700 @ 40/15+0.05
4529 - 4385 - 4786
ELO 3.65±4.69

LTC no regression
10590 @ 40/60+0.05
3181 - 3042 - 4367
ELO 4.56±5.07

+++ zurichess/uci.go
@@ -262,7 +262,7 @@ func (uci *UCI) go_(line string) error {
 		case "movestogo":
 			i++
 			t, _ := strconv.Atoi(args[i])
-			uci.timeControl.MovesToGo = t
+			uci.timeControl.MovesToGo = int32(t)
 		case "movetime":
 			i++
 			t, _ := strconv.Atoi(args[i])
